design.py: use Bridge DP
Share an implementation among multiple objects, and this fact should be
hidden from the client.

Decoupling interface and implementation. This decoupling encourages
layering that can lead to a better-structured system. The high level
part of a system only has to know about Abstraction and implementor.
Improved extensibility. Extend the Abstraction and Implementor
hierarchies independently.