design.py: change Bridge DP to dependency injection
There is a problem with last version of design.py, in fact, Bridge DP
is not suitable for this script. In Bridge, there are more than one
interface, but this is not necessary for design.py. Because there is no
requirement on interface, so read and write methods can be contained in
just one class. Then dependency injection would supply extensibility.